+++ server/BUILD.bazel
@@ -49,13 +49,13 @@ go_library(
         "//privilege/privileges/ldap",
         "//server/constvar",
         "//server/err",
-        "//server/internal",
         "//server/internal/column",
         "//server/internal/dump",
         "//server/internal/handshake",
         "//server/internal/httputil",
         "//server/internal/info",
         "//server/internal/parse",
+        "//server/internal/pocketio",
         "//server/internal/resultset",
         "//server/internal/util",
         "//server/metrics",
@@ -169,10 +169,11 @@ go_test(
         "//parser/charset",
         "//parser/model",
         "//parser/mysql",
-        "//server/internal",
+        "//server/internal/bytesconn",
         "//server/internal/column",
         "//server/internal/handshake",
         "//server/internal/parse",
+        "//server/internal/pocketio",
         "//server/internal/resultset",
         "//server/internal/testutil",
         "//server/internal/util",
