--- server/internal/BUILD.bazel
@@ -1,34 +1,0 @@
-load("@io_bazel_rules_go//go:def.bzl", "go_library", "go_test")
-
-go_library(
-    name = "internal",
-    srcs = ["packetio.go"],
-    importpath = "github.com/pingcap/tidb/server/internal",
-    visibility = ["//server:__subpackages__"],
-    deps = [
-        "//parser/mysql",
-        "//parser/terror",
-        "//server/err",
-        "//server/internal/util",
-        "//server/metrics",
-        "//sessionctx/variable",
-        "@com_github_klauspost_compress//zstd",
-        "@com_github_pingcap_errors//:errors",
-    ],
-)
-
-go_test(
-    name = "internal_test",
-    timeout = "short",
-    srcs = ["packetio_test.go"],
-    embed = [":internal"],
-    flaky = True,
-    shard_count = 4,
-    deps = [
-        "//parser/mysql",
-        "//server/internal/testutil",
-        "//server/internal/util",
-        "@com_github_klauspost_compress//zstd",
-        "@com_github_stretchr_testify//require",
-    ],
-)
