+++ server/internal/bytesconn/BUILD.bazel
@@ -1,0 +1,8 @@
+load("@io_bazel_rules_go//go:def.bzl", "go_library")
+
+go_library(
+    name = "bytesconn",
+    srcs = ["testutil.go"],
+    importpath = "github.com/pingcap/tidb/server/internal/bytesconn",
+    visibility = ["//server:__subpackages__"],
+)
