+++ server/internal/pocketio/BUILD.bazel
@@ -1,0 +1,34 @@
+load("@io_bazel_rules_go//go:def.bzl", "go_library", "go_test")
+
+go_library(
+    name = "pocketio",
+    srcs = ["packetio.go"],
+    importpath = "github.com/pingcap/tidb/server/internal/pocketio",
+    visibility = ["//server:__subpackages__"],
+    deps = [
+        "//parser/mysql",
+        "//parser/terror",
+        "//server/err",
+        "//server/internal/util",
+        "//server/metrics",
+        "//sessionctx/variable",
+        "@com_github_klauspost_compress//zstd",
+        "@com_github_pingcap_errors//:errors",
+    ],
+)
+
+go_test(
+    name = "pocketio_test",
+    timeout = "short",
+    srcs = ["packetio_test.go"],
+    embed = [":pocketio"],
+    flaky = True,
+    shard_count = 4,
+    deps = [
+        "//parser/mysql",
+        "//server/internal/bytesconn",
+        "//server/internal/util",
+        "@com_github_klauspost_compress//zstd",
+        "@com_github_stretchr_testify//require",
+    ],
+)
