+++ server/internal/pocketio/packetio.go
@@ -33,7 +33,7 @@
 // The above copyright notice and this permission notice shall be included in all
 // copies or substantial portions of the Software.
 
-package internal
+package pocketio
 
 import (
 	"bufio"
